+++ NowPlayingTweet/AppDelegate.swift
@@ -7,7 +7,6 @@
 
 import Cocoa
 import SwifterMac
-import ScriptingUtilities
 import iTunesScripting
 
 @NSApplicationMain
@@ -18,10 +17,10 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     let statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.variableLength)
 
     let userDefaults: UserDefaults = UserDefaults.standard
-    
+
     let twitterAccount: TwitterAccount = TwitterAccount()
-    
-    var playerInfo: iTunesPlayerInfo?
+
+    var playerInfo: iTunesPlayerInfo = iTunesPlayerInfo()
 
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         // Insert code here to initialize your application
@@ -36,8 +35,11 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             ]
         self.userDefaults.register(defaults: defaultSettings)
 
-        self.statusItem.title = "NPT"
-        self.statusItem.highlightMode = true
+        if let button = self.statusItem.button {
+            let image = NSImage(named: NSImage.Name("StatusBarItem"))
+            image?.isTemplate = true
+            button.image = image
+        }
         self.statusItem.menu = self.menu
 
         if self.userDefaults.bool(forKey: "AutoTweet") {
@@ -61,23 +63,19 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             return
         }
 
-        self.playerInfo = self.getPlayerInfo()
         self.postTweet()
     }
 
-    func getPlayerInfo() -> iTunesPlayerInfo {
-        let iTunes: iTunesApplication = ScriptingUtilities.application(name: "iTunes") as! iTunesApplication
-        //let iTunes = ScriptingUtilities.application(bundleIdentifier: "com.apple.iTunes") as! iTunesApplication
-
-        let currentTrack: iTunesTrack = iTunes.currentTrack!
-
-        return iTunesPlayerInfo(currentTrack)
+    @IBAction func tweetNowPlaying(_ sender: Any) {
+        self.postTweet()
     }
 
     func postTweet() {
+        self.playerInfo.updateTrack()
+
         let tweetText = self.createTweetText()
         if self.userDefaults.bool(forKey: "TweetWithImage") {
-            self.twitterAccount.tweet(text: tweetText, with: self.playerInfo?.artwork)
+            self.twitterAccount.tweet(text: tweetText, with: self.playerInfo.artwork)
         } else {
             self.twitterAccount.tweet(text: tweetText)
         }
@@ -86,33 +84,18 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     func createTweetText() -> String {
         var format = self.userDefaults.string(forKey: "TweetFormat")!
 
-        while true {
-            if let range = format.range(of: "{{Title}}") {
-                format.replaceSubrange(range, with: (self.playerInfo?.title)!)
-                continue
-            }
+        let convertDictionary: [String : String] = [
+            "{{Title}}" : self.playerInfo.title!,
+            "{{Artist}}" : self.playerInfo.artist!,
+            "{{Album}}" : self.playerInfo.album!,
+            "{{AlbumArtist}}" : self.playerInfo.albumArtist!,
+            "{{BitRate}}" : String(self.playerInfo.bitRate!),
+        ]
 
-            if let range = format.range(of: "{{Artist}}") {
-                format.replaceSubrange(range, with: (self.playerInfo?.artist)!)
-                continue
+        for (from,to) in convertDictionary {
+            while let range = format.range(of: from) {
+                format.replaceSubrange(range, with: to)
             }
-
-            if let range = format.range(of: "{{Album}}") {
-                format.replaceSubrange(range, with: (self.playerInfo?.album)!)
-                continue
-            }
-
-            if let range = format.range(of: "{{AlbumArtist}}") {
-                format.replaceSubrange(range, with: (self.playerInfo?.albumArtist)!)
-                continue
-            }
-
-            if let range = format.range(of: "{{BitRate}}") {
-                format.replaceSubrange(range, with: String((self.playerInfo?.bitRate)!))
-                continue
-            }
-
-            break
         }
 
         return format