+++ NowPlayingTweet/NSStoryboard++.swift
@@ -1,0 +1,16 @@
+/**
+ *  NSStoryboard++.swift
+ *  NowPlayingTweet
+ *
+ *  © 2018 kPherox.
+**/
+
+import Foundation
+import AppKit
+
+extension NSStoryboard.SceneIdentifier {
+    static let generalViewController = NSStoryboard.SceneIdentifier("GeneralViewController")
+    static let advancedViewController = NSStoryboard.SceneIdentifier("AdvancedViewController")
+    static let accountsViewController = NSStoryboard.SceneIdentifier("AccountsViewController")
+
+}