+++ NowPlayingTweet/NSToolbarItem++.swift
@@ -1,0 +1,15 @@
+/**
+ *  NSToolbarItem++.swift
+ *  NowPlayingTweet
+ *
+ *  © 2018 kPherox.
+**/
+
+import Foundation
+import AppKit
+
+extension NSToolbarItem.Identifier {
+    static let general =  NSToolbarItem.Identifier("General")
+    static let advanced =  NSToolbarItem.Identifier("Advanced")
+    static let accounts =  NSToolbarItem.Identifier("Accounts")
+}