+++ NowPlayingTweet/PreferencesWindowController.swift
@@ -9,35 +9,36 @@ import Cocoa
 
 class PreferencesWindowController: NSWindowController {
 
+    @IBOutlet weak var toolbar: NSToolbar!
     @IBOutlet weak var generalToolbarItem: NSToolbarItem!
     @IBOutlet weak var accountsToolbarItem: NSToolbarItem!
     @IBOutlet weak var advancedToolbarItem: NSToolbarItem!
 
     override func windowDidLoad() {
         super.windowDidLoad()
+
+        self.toolbar.selectedItemIdentifier = self.generalToolbarItem.itemIdentifier
     }
 
     @IBAction func switchViewController(_ sender: NSToolbarItem) {
-        switch sender.itemIdentifier.rawValue {
-        case "General":
-            self.replaceViewController("GeneralViewController")
-        case "Accounts":
-            self.replaceViewController("AccountsViewController")
-        case "Advanced":
-            self.replaceViewController("AdvancedViewController")
+        switch sender.itemIdentifier {
+        case .general:
+            self.replaceViewController(identifier: .generalViewController)
+        case .accounts:
+            self.replaceViewController(identifier: .accountsViewController)
+        case .advanced:
+            self.replaceViewController(identifier: .advancedViewController)
         default:
             break
         }
     }
 
-    private func replaceViewController(_ identifier: String) {
-        let storyboard = NSStoryboard(name: NSStoryboard.Name(rawValue: "Main"), bundle: nil)
-
-        if let viewController = (storyboard.instantiateController(withIdentifier: NSStoryboard.SceneIdentifier(rawValue: identifier)) as? NSViewController) {
-            let windowFrame = self.window?.frame
+    private func replaceViewController(identifier: NSStoryboard.SceneIdentifier) {
+        if let viewController = self.storyboard?.instantiateController(withIdentifier: identifier) as? NSViewController {
+            let windowFrame: NSRect = (self.window?.frame)!
             var newWindowFrame: NSRect = (self.window?.frameRect(forContentRect: viewController.view.frame))!
-            newWindowFrame.origin.x = (windowFrame?.origin.x)!
-            newWindowFrame.origin.y = (windowFrame?.origin.y)! + (windowFrame?.size.height)! - (newWindowFrame.size.height)
+            newWindowFrame.origin.x = windowFrame.origin.x
+            newWindowFrame.origin.y = windowFrame.origin.y + windowFrame.size.height - newWindowFrame.size.height
 
             self.window?.contentViewController = nil
             self.window?.setFrame(newWindowFrame, display: true, animate: true)
@@ -46,4 +47,5 @@ class PreferencesWindowController: NSWindowController {
             return
         }
     }
+
 }